+++ package.json
@@ -39,12 +39,12 @@
     "typings/index.d.ts"
   ],
   "dependencies": {
-    "axios": "^1.6.8"
+    "axios": "^1.6.8",
+    "form-data": "^4.0.0"
   },
   "devDependencies": {
     "chai": "^4.4.1",
     "docdash": "^2.0.2",
-    "form-data": "^4.0.0",
     "glob": "^10.3.12",
     "jsdoc": "^4.0.2",
     "jsdoc-to-markdown": "^8.0.1",
