+++ php/.htaccess
