+++ php/tokens.php
@@ -1,0 +1,3 @@
+<?php
+
+$db_tokens = array("my_secret_token_here", "a_different_token_here");
