--- src/tokens.php
@@ -1,7 +1,0 @@
-<?php
-
-$db_tokens_map = array(
-    'CompliBot' => "KGKbkjJKKLMhnkJjkbjkk"
-);
-
-$db_tokens = array_values($db_tokens_map);
